ADDED: If Inch Is Null It Should Return False

+++ src/main/java/com/bl/demo/model/Feet.java
@@ -2,7 +2,7 @@ package com.bl.demo.model;
 
 public class Feet extends Units {
 
-    public Feet(int length) {
+    public Feet(double length) {
         super(length);
     }
 }
